pyproject.toml
# Copyright 2026 Enactic, Inc.
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#     http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

[project]
name = "dora-openarm-docker-policy-server"
version = "1.0.0"
authors = [{ name = "Enactic, Inc." }]
description = "Dora node to communicate with Docker based policy server"
license = { text = "Apache-2.0" }
readme = "README.md"
requires-python = ">=3.10"

dependencies = ["dora-rs >= 0.3.9"]

[dependency-groups]
dev = [
  "pytest >= 8.1.1",
  "ruff == 0.14.11", # sync with .pre-commit-config.yaml
]

[project.scripts]
dora-openarm-docker-policy-server = "dora_openarm_docker_policy_server.main:main"

[tool.ruff.lint]
extend-select = [
  "D",   # pydocstyle
  "UP"
]
